+++ src/lib/core/point.cpp
@@ -9,12 +9,10 @@ namespace retronomicon::lib::core{
     /**
      * @brief constructor with 2 points 
      * 
-     * Notes : still considering wether we should use int or float or doubles
-     * 
      * @param x the x position
      * @param y the y position
      */
-    Point::Point(int x, int y){
+    Point::Point(float x, float y){
         set(x,y);
     }
 
@@ -24,7 +22,7 @@ namespace retronomicon::lib::core{
      * @param x the x position
      * @param y the y position
      */
-    void Point::set(int x,int y){
+    void Point::set(float x,float y){
         m_x = x;
         m_y = y;
     }
@@ -34,7 +32,7 @@ namespace retronomicon::lib::core{
      * 
      * @return the x position of the point
      */
-    int Point::getX() const {
+    float Point::getX() const {
         return m_x;
     }
 
@@ -43,7 +41,7 @@ namespace retronomicon::lib::core{
      * 
      * @return the y position of the point
      */
-    int Point::getY() const {
+    float Point::getY() const {
         return m_y;
     }
 
@@ -54,8 +52,8 @@ namespace retronomicon::lib::core{
      * @return the distance in float
      */
     float Point::calculateDistance(Point* target){
-        float distance_x = abs(((float) m_x - (float) target->getX()));
-        float distance_y = abs(((float) m_y - (float) target->getY()));
+        float distance_x = abs((m_x - target->getX()));
+        float distance_y = abs((m_y - target->getY()));
         return sqrt((distance_x*distance_x) + (distance_y*distance_y));
     }
 }
